requirements.txt
django==2.0.3
mysqlclient==1.3.12
djangorestframework
pygments
